- Rename Customers into Users - Added OpenApi documentation to endpoints - Cleaned up ambiguous endpoint paths - Added Specials feature implementation

+++ src/main/java/za/co/serengti/receipts/dto/TillDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class TillDTO {
-    private short tillNumber;
+    private Long tillNumber;
 }
